+++ framework/registerCli.php
@@ -1,0 +1,6 @@
+<?php
+
+use Framework\Cli\Cli;
+use Framework\Test\Cli\TestRunCommand;
+
+Cli::registerCommand(new TestRunCommand());