+++ tests/Unit/TestTest.php
@@ -1,0 +1,16 @@
+<?php
+
+use \Framework\Test\TestCase;
+
+class TestTest extends TestCase
+{
+    public function testCase1()
+    {
+        $this->assertTrue(true);
+    }
+
+    public function testCase2()
+    {
+        $this->assertTrue(true);
+    }
+}